#78 storage hook OnDisconnect only execute the relevant code if the client has been reconnected

+++ hooks/storage/badger/badger.go
@@ -182,9 +182,11 @@ func (h *Hook) OnDisconnect(cl *mqtt.Client, _ error, expire bool) {
 		return
 	}
 
-	err := h.db.Delete(clientKey(cl), new(storage.Client))
-	if err != nil {
-		h.Log.Error().Err(err).Interface("data", clientKey(cl)).Msg("failed to delete client data")
+	if cl.StopCause() != packets.ErrSessionTakenOver {
+		err := h.db.Delete(clientKey(cl), new(storage.Client))
+		if err != nil {
+			h.Log.Error().Err(err).Interface("data", clientKey(cl)).Msg("failed to delete client data")
+		}
 	}
 }
